Soporte para interfaz de fichero. [API Murano]

+++ ginn/api/tests/murano_tests.py
@@ -5,6 +5,7 @@ from __future__ import print_function
 
 ERRCODENOTFOUND = 1
 ERRCODENOTIMPLEMENTED = 2
+ERRFILENOTFOUND = 3
 
 import argparse
 import sys, os
@@ -27,15 +28,18 @@ def prueba_bala(codigo = None):
         b.articulo.productoVenta.descripcion,
         b.puid))
     murano.create_bala(b)
-    for c in b.articulo.parteDeProduccion.consumos:
-        if c.silo:
-            print("Consumiendo %f de %s (%s)..." % (c.cantidad,
-                c.productoCompra.descripcion,
-                c.silo.nombre))
-        else:
-            print("Consumiendo %f de %s..." % (c.cantidad,
-                                               c.productoCompra.descripcion))
-        murano.consumir(c.productoCompra, c.cantidad, consumo = c)
+    if b.articulo.parteDeProduccion:
+        # TODO: PORASQUI. ¿Se podría calcular la parte proporcional, o es demasiado para unas pruebas?
+        print("WARNING: Considerando consumos del parte de producción completo.")
+        for c in b.articulo.parteDeProduccion.consumos:
+            if c.silo:
+                print("Consumiendo %f de %s (%s)..." % (c.cantidad,
+                    c.productoCompra.descripcion,
+                    c.silo.nombre))
+            else:
+                print("Consumiendo %f de %s..." % (c.cantidad,
+                                                   c.productoCompra.descripcion))
+            murano.consumir(c.productoCompra, c.cantidad, consumo = c)
 
 def prueba_rollo(codigo = None):
     if not codigo:
@@ -46,21 +50,23 @@ def prueba_rollo(codigo = None):
         r.articulo.productoVenta.descripcion,
         r.puid))
     murano.create_rollo(r)
-    for c in r.articulo.parteDeProduccion.consumos:
-        print("Consumiendo %f de %s..." % (c.cantidad,
-                                           c.productoCompra.descripcion))
-        murano.consumir(c.productoCompra, c.cantidad)
-    # Para probar, consumiré la partida de carga completa:
-    if r.articulo.parteDeProduccion.partidaCarga:
-        for b in r.articulo.parteDeProduccion.partidaCarga.balas:
-            # El almacén donde estaba la bala **antes** de consumirla está 
-            # en Murano. En ginn el almacén es None. Confiamos en Murano
-            # (mediante la función delete_articulo del módulo murano)
-            # para buscar e indicar de qué almacén debe consumir la bala. Sería
-            # muy complicado (y más lento) determinarlo en ginn antes de lanzar
-            # el consumo a la pasarela.
-            print("Consumiendo %s (%s)..." % (b.codigo, b.puid))
-            murano.delete_articulo(b.articulo)
+    if r.articulo.parteDeProduccion:
+        print("WARNING: Considerando consumos del parte de producción completo.")
+        for c in r.articulo.parteDeProduccion.consumos:
+            print("Consumiendo %f de %s..." % (c.cantidad,
+                                               c.productoCompra.descripcion))
+            murano.consumir(c.productoCompra, c.cantidad)
+        # Para probar, consumiré la partida de carga completa:
+        if r.articulo.parteDeProduccion.partidaCarga:
+            for b in r.articulo.parteDeProduccion.partidaCarga.balas:
+                # El almacén donde estaba la bala **antes** de consumirla está 
+                # en Murano. En ginn el almacén es None. Confiamos en Murano
+                # (mediante la función delete_articulo del módulo murano)
+                # para buscar e indicar de qué almacén debe consumir la bala. Sería
+                # muy complicado (y más lento) determinarlo en ginn antes de lanzar
+                # el consumo a la pasarela.
+                print("Consumiendo %s (%s)..." % (b.codigo, b.puid))
+                murano.delete_articulo(b.articulo)
 
 def prueba_pale(codigo = None):
     if not codigo:
@@ -71,14 +77,16 @@ def prueba_pale(codigo = None):
         p.productoVenta.descripcion,
         p.puid))
     murano.create_pale(p)
-    for c in p.parteDeProduccion.consumos:
-        print("Consumiendo %f de %s..." % (c.cantidad,
-                                           c.productoCompra.descripcion))
-        murano.consumir(c.productoCompra, c.cantidad)
-    # Consumo los bigbags de fibra empleados en rellenar las bolsas
-    for bb in p.parteDeProduccion.bigbags:
-        print("Consumiendo %s (%s)..." % (bb.codigo, bb.puid))
-        murano.delete_articulo(bb.articulo)
+    if p.parteDeProduccion:
+        print("WARNING: Considerando consumos del parte de producción completo.")
+        for c in p.parteDeProduccion.consumos:
+            print("Consumiendo %f de %s..." % (c.cantidad,
+                                               c.productoCompra.descripcion))
+            murano.consumir(c.productoCompra, c.cantidad)
+        # Consumo los bigbags de fibra empleados en rellenar las bolsas
+        for bb in p.parteDeProduccion.bigbags:
+            print("Consumiendo %s (%s)..." % (bb.codigo, bb.puid))
+            murano.delete_articulo(bb.articulo)
 
 def prueba_bigbag(codigo = None):
     if not codigo:
@@ -89,19 +97,21 @@ def prueba_bigbag(codigo = None):
         bb.articulo.productoVenta.descripcion,
         bb.puid))
     murano.create_bigbag(bb)
-    for c in bb.articulo.parteDeProduccion.consumos:
-        if c.silo:
-            print("Consumiendo %f de %s (%s)..." % (c.cantidad,
-                c.productoCompra.descripcion,
-                c.silo.nombre))
-        else:
-            print("Consumiendo %f de %s..." % (c.cantidad,
-                                               c.productoCompra.descripcion))
-        murano.consumir(c.productoCompra, c.cantidad, consumo = c)
+    if bb.articulo.parteDeProduccion:
+        print("WARNING: Considerando consumos del parte de producción completo.")
+        for c in bb.articulo.parteDeProduccion.consumos:
+            if c.silo:
+                print("Consumiendo %f de %s (%s)..." % (c.cantidad,
+                    c.productoCompra.descripcion,
+                    c.silo.nombre))
+            else:
+                print("Consumiendo %f de %s..." % (c.cantidad,
+                                                   c.productoCompra.descripcion))
+            murano.consumir(c.productoCompra, c.cantidad, consumo = c)
     # Los partes de reembolsado se ignoran. Solo la fibra fabricada
     # directamente para almacenar en bigbags en lugar de en balas.
 
-def prueba_codigo(codigo):
+def prueba_codigo(codigo, consumir = False):
     mapping = {pclases.PREFIJO_ROLLO: pclases.Rollo,
                pclases.PREFIJO_BALA: pclases.Bala,
                pclases.PREFIJO_BIGBAG: pclases.Bigbag,
@@ -115,7 +125,7 @@ def prueba_codigo(codigo):
         if codigo.startswith(prefijo):
             clase_pclases = mapping[prefijo]
             objeto = clase_pclases.selectBy(codigo = codigo)[0]
-            prueba_objeto(objeto)
+            prueba_objeto(objeto, consumir)
     if not objeto:
         if codigo.startswith(pclases.PREFIJO_BOLSA):
             print("Código de bolsa detectado. Debe insertar al menos una "
@@ -140,27 +150,57 @@ def prueba_codigo(codigo):
             print("El código %s no se reconoce.")
             sys.exit(ERRCODENOTFOUND)
 
-def prueba_objeto(objeto):
+def prueba_objeto(objeto, consumir = False):
     if isinstance(objeto, (pclases.Rollo,
                            pclases.RolloDefectuoso,
                            pclases.RolloC)):
-        print("Insertando rollo %s (%s)..." % (objeto.codigo, objeto.puid))
-        murano.create_rollo(objeto)
+        if not consumir:
+            print("Insertando rollo %s (%s)..." % (objeto.codigo, objeto.puid))
+            murano.create_rollo(objeto)
+        else:
+            prueba_rollo(objeto.codigo)
     elif isinstance(objeto, (pclases.Bala,
                              pclases.BalaCable)):
-        print("Insertando bala %s (%s)..." % (objeto.codigo, objeto.puid))
-        murano.create_bala(objeto)
+        if not consumir:
+            print("Insertando bala %s (%s)..." % (objeto.codigo, objeto.puid))
+            murano.create_bala(objeto)
+        else:
+            prueba_bala(objeto.codigo)
     elif isinstance(objeto, pclases.Bigbag):
-        print("Insertando bigbag %s (%s)..." % (objeto.codigo, objeto.puid))
-        murano.create_bigbag(objeto)
+        if not consumir:
+            print("Insertando bigbag %s (%s)..." % (objeto.codigo, objeto.puid))
+            murano.create_bigbag(objeto)
+        else:
+            prueba_bigbag(objeto.codigo)
     elif isinstance(objeto, pclases.Caja):
-        print("Insertando caja %s (%s)..." % (objeto.codigo, objeto.puid))
-        murano.create_caja(objeto)
+        if not consumir:
+            print("Insertando caja %s (%s)..." % (objeto.codigo, objeto.puid))
+            murano.create_caja(objeto)
+        else:
+            prueba_caja(objeto.codigo)
     elif isinstance(objeto, pclases.Pale):
-        print("Insertando palé %s (%s)..." % (objeto.codigo, objeto.puid))
-        murano.create_pale(objeto)
+        if not consumir:
+            print("Insertando palé %s (%s)..." % (objeto.codigo, objeto.puid))
+            murano.create_pale(objeto)
+        else:
+            prueba_pale(objeto.codigo)
     else:
         raise NotImplementedError, "%s no soportado" % (objeto.puid)
+
+def parse_file(fsource):
+    """
+    Función generadora. Lee los códigos a insertar de un fichero de texto.
+    Devuelve un código en cada iteración.
+    """
+    try:
+        f = open(fsource)
+    except IOError:
+        print("El fichero %s no existe." % (fsource))
+        sys.exit(ERRFILENOTFOUND)
+    else:
+        for l in f.readlines():
+            for codigo in l.strip().upper().split():
+                yield codigo
 
 def main():
     ## Parámetros
@@ -178,10 +218,14 @@ def main():
     parser.add_argument("-g", "--bigbags", dest = "bigbags",
             help = "Inserta los últimos n bigbags y sus consumos.",
             default = 0, type = int)
-    parser.add_argument("-o", "--codigo", dest = "codigo",
+    parser.add_argument("-o", "--codigo", dest = "codigo", nargs = "+",
             help = "Especifica el código del artículo a insertar. "
-                   "**No tiene en cuenta los consumos relacionados.**",
+                   "**No tiene en cuenta los consumos.**", 
             default = None, type = str)
+    parser.add_argument("-f", "--file_source", dest = "file_source",
+            help = "Lee los códigos a insertar desde un fichero de texto."
+                   "**Realiza también los consumos relacionados.**",
+            type = str)
     if len(sys.argv) == 1:
         parser.print_help()
     args = parser.parse_args()
@@ -203,7 +247,11 @@ def main():
         for bigbag in bigbags.limit(args.bigbags):
             prueba_bigbag(bigbag.codigo)
     if args.codigo:
-        prueba_codigo(args.codigo)
+        for codigo in args.codigo:
+            prueba_codigo(codigo)
+    if args.file_source: 
+        for codigo in parse_file(args.file_source):
+            prueba_codigo(codigo, consumir = True)
 
 if __name__ == "__main__":
     main()
